feat: export computeTargetBalances and prod-network for yds consumption
