+++ packages/portfolio-api/README.md
@@ -1,3 +1,26 @@
 # Portfolio management API
 
 API shared between on-chain contract and external clients
+
+## YDS Target Balances
+
+YDS should import the shared target-balance helper and production network data
+from this public package:
+
+```js
+import { computeTargetBalances } from '@agoric/portfolio-api/src/target-balances.js';
+import { PROD_NETWORK } from '@agoric/portfolio-api/src/network/prod-network.js';
+
+const changedTargets = computeTargetBalances({
+  brand: usdcBrand,
+  currentBalances,
+  balanceDelta,
+  targetAllocation,
+  network: PROD_NETWORK,
+  depositFromChain,
+});
+```
+
+`changedTargets` contains only balances that need to change. Use signed
+minor-unit deltas: positive for deposits, negative for withdrawals, and `0n` for
+rebalances.
