+++ packages/portfolio-api/src/main.js
@@ -1,6 +1,8 @@
 export * from './constants.js';
 export * from './instruments.js';
+export * from './network/prod-network.ts';
 export * from './resolver.js';
+export * from './target-balances.ts';
 export * from './type-guards.ts';
 
 // eslint-disable-next-line import/export -- just types
