+++ packages/portfolio-api/src/types-index.d.ts
@@ -1,1 +1,3 @@
+export type * from './network/network-spec.js';
+export type * from './target-balances.js';
 export type * from './types.js';
