+++ packages/portfolio-api/test/snapshots/exports.test.ts.md
@@ -12,18 +12,22 @@ Generated by [AVA](https://avajs.dev).
       'ACCOUNT_DUST_EPSILON',
       'AxelarChain',
       'CaipChainIds',
+      'DEFAULT_DELTA_SOFT_MIN',
       'DEFAULT_FLOW_CONFIG',
       'Eip155ChainIds',
       'EvmWalletOperationType',
       'FlowConfigShape',
       'FlowFeaturesShape',
       'InstrumentId',
+      'PROD_NETWORK',
       'RebalanceStrategy',
       'SupportedChain',
+      'TargetBalanceError',
       'TxStatus',
       'TxType',
       'UsdcTokenIds',
       'YieldProtocol',
+      'computeTargetBalances',
       'isBeefyInstrumentId',
       'isDepositFromChainRef',
       'isERC4626InstrumentId',
