+++ packages/portfolio-contract/README.md
@@ -32,3 +32,29 @@ For details on making offers and querying vstorage, see
  - `src/type-guards.ts` - types and pattern guards
  - `test/portfolio-agents.ts` - example client code
  - `test/snapshots/*.md` - example vstorage data
+
+## Off-chain Target Balance API
+
+YDS should consume the shared target-balance helper from the public
+`@agoric/portfolio-api` package. That package participates in the repo's
+after-merge canary publish flow, so production network changes can propagate
+without waiting for `@aglocal/portfolio-contract` packaging.
+
+```js
+import { computeTargetBalances } from '@agoric/portfolio-api/src/target-balances.js';
+import { PROD_NETWORK } from '@agoric/portfolio-api/src/network/prod-network.js';
+
+const changedTargets = computeTargetBalances({
+  brand: usdcBrand,
+  currentBalances,
+  balanceDelta,
+  targetAllocation,
+  network: PROD_NETWORK,
+  depositFromChain,
+});
+```
+
+`changedTargets` contains only balances that need to change. YDS should pass the
+same USDC brand used by its `currentBalances`, use signed minor-unit deltas
+(positive deposits, negative withdrawals), and keep importing the `.js` subpaths
+so the same source works after packaging.
