+++ packages/portfolio-contract/tools/network/prod-network.ts
@@ -1,387 +1,4 @@
-import type { NetworkSpec } from './network-spec.js';
+import { PROD_NETWORK } from '@agoric/portfolio-api/src/network/prod-network.js';
 
-// Initial production network in NetworkSpec format
-export const PROD_NETWORK: NetworkSpec = harden({
-  // Enable debug diagnostics to aid troubleshooting in tests
-  debug: true,
-  environment: 'prod',
-  chains: [
-    { name: 'agoric', control: 'local' },
-    { name: 'noble', control: 'ibc' },
-    { name: 'Arbitrum', control: 'axelar' },
-    { name: 'Avalanche', control: 'axelar' },
-    { name: 'Base', control: 'axelar' },
-    { name: 'Ethereum', control: 'axelar' },
-    { name: 'Optimism', control: 'axelar' },
-  ],
-  pools: [
-    { pool: 'Aave_Arbitrum', chain: 'Arbitrum', protocol: 'Aave' },
-    { pool: 'Aave_Avalanche', chain: 'Avalanche', protocol: 'Aave' },
-    { pool: 'Aave_Base', chain: 'Base', protocol: 'Aave' },
-    { pool: 'Aave_Ethereum', chain: 'Ethereum', protocol: 'Aave' },
-    { pool: 'Aave_Optimism', chain: 'Optimism', protocol: 'Aave' },
-    { pool: 'Beefy_re7_Avalanche', chain: 'Avalanche', protocol: 'Beefy' },
-    {
-      pool: 'Beefy_morphoGauntletUsdc_Ethereum',
-      chain: 'Ethereum',
-      protocol: 'Beefy',
-    },
-    {
-      pool: 'Beefy_morphoSmokehouseUsdc_Ethereum',
-      chain: 'Ethereum',
-      protocol: 'Beefy',
-    },
-    { pool: 'Beefy_morphoSeamlessUsdc_Base', chain: 'Base', protocol: 'Beefy' },
-    {
-      pool: 'Beefy_compoundUsdc_Optimism',
-      chain: 'Optimism',
-      protocol: 'Beefy',
-    },
-    {
-      pool: 'Beefy_compoundUsdc_Arbitrum',
-      chain: 'Arbitrum',
-      protocol: 'Beefy',
-    },
-    {
-      pool: 'ERC4626_morphoClearstarHighYieldUsdc_Ethereum',
-      chain: 'Ethereum',
-      protocol: 'ERC4626',
-      blockDepositReason: 'AT_CAPACITY',
-      blockWithdrawReason: 'LOW_LIQUIDITY',
-    },
-    {
-      pool: 'ERC4626_morphoSeamlessUsdcVault_Base',
-      chain: 'Base',
-      protocol: 'ERC4626',
-      blockDepositReason: 'AT_CAPACITY',
-    },
-    { pool: 'Compound_Arbitrum', chain: 'Arbitrum', protocol: 'Compound' },
-    { pool: 'Compound_Base', chain: 'Base', protocol: 'Compound' },
-    { pool: 'Compound_Ethereum', chain: 'Ethereum', protocol: 'Compound' },
-    { pool: 'Compound_Optimism', chain: 'Optimism', protocol: 'Compound' },
-    { pool: 'USDN', chain: 'noble', protocol: 'USDN' },
-    { pool: 'USDNVault', chain: 'noble', protocol: 'USDN' },
-  ],
-  localPlaces: [
-    { id: '<Deposit>', chain: 'agoric' },
-    { id: '<Cash>', chain: 'agoric' },
-    { id: '+agoric', chain: 'agoric' },
-  ],
-  links: [
-    // USDN costs a fee to get into
-    {
-      src: '@noble',
-      dest: 'USDN',
-      transfer: 'local',
-      variableFeeBps: 10,
-      timeSec: 0,
-      feeMode: 'toUSDN',
-    },
-    {
-      src: '@noble',
-      dest: 'USDNVault',
-      transfer: 'local',
-      variableFeeBps: 10,
-      timeSec: 0,
-      feeMode: 'toUSDN',
-    },
-
-    // CCTP slow (inbound auto-forward compressed: EVM -> @agoric)
-    // Latency kept at 1080s (assuming IBC forward overlaps); adjust if sequential.
-    {
-      src: '@Arbitrum',
-      dest: '@agoric',
-      transfer: 'cctpToNoble',
-      variableFeeBps: 0,
-      timeSec: 1080,
-      feeMode: 'evmToNoble',
-    },
-    {
-      src: '@Avalanche',
-      dest: '@agoric',
-      transfer: 'cctpToNoble',
-      variableFeeBps: 0,
-      timeSec: 30, // Avalanche has instant finality (~7 seconds observed)
-      feeMode: 'evmToNoble',
-    },
-    {
-      src: '@Ethereum',
-      dest: '@agoric',
-      transfer: 'cctpToNoble',
-      variableFeeBps: 0,
-      timeSec: 1080,
-      feeMode: 'evmToNoble',
-    },
-    {
-      src: '@Optimism',
-      dest: '@agoric',
-      transfer: 'cctpToNoble',
-      variableFeeBps: 0,
-      timeSec: 1080,
-      feeMode: 'evmToNoble',
-    },
-    {
-      src: '@Base',
-      dest: '@agoric',
-      transfer: 'cctpToNoble',
-      variableFeeBps: 0,
-      timeSec: 1080,
-      feeMode: 'evmToNoble',
-    },
-    // CCTP return (fast on return path); 1 USDC minimum
-    {
-      src: '@noble',
-      dest: '@Arbitrum',
-      transfer: 'cctpFromNoble',
-      variableFeeBps: 0,
-      timeSec: 20,
-      min: 1_000_000n,
-      feeMode: 'makeEvmAccount',
-    },
-    {
-      src: '@noble',
-      dest: '@Avalanche',
-      transfer: 'cctpFromNoble',
-      variableFeeBps: 0,
-      timeSec: 20,
-      min: 1_000_000n,
-      feeMode: 'makeEvmAccount',
-    },
-    {
-      src: '@noble',
-      dest: '@Ethereum',
-      transfer: 'cctpFromNoble',
-      variableFeeBps: 0,
-      timeSec: 20,
-      min: 1_000_000n,
-      feeMode: 'makeEvmAccount',
-    },
-    {
-      src: '@noble',
-      dest: '@Optimism',
-      transfer: 'cctpFromNoble',
-      variableFeeBps: 0,
-      timeSec: 20,
-      min: 1_000_000n,
-      feeMode: 'makeEvmAccount',
-    },
-    {
-      src: '@noble',
-      dest: '@Base',
-      transfer: 'cctpFromNoble',
-      variableFeeBps: 0,
-      timeSec: 20,
-      min: 1_000_000n,
-      feeMode: 'makeEvmAccount',
-    },
-    // IBC connectivity (explicit both directions required for USDN and other noble-origin flows)
-    {
-      src: '@agoric',
-      dest: '@noble',
-      transfer: 'ibc',
-      variableFeeBps: 0,
-      timeSec: 10,
-    },
-    {
-      src: '@noble',
-      dest: '@agoric',
-      transfer: 'ibc',
-      variableFeeBps: 0,
-      timeSec: 10,
-    },
-
-    // CCTPv2 direct EVM-to-EVM routes (full mesh connectivity)
-    // Estimated time: ~13-60 seconds depending on finality threshold
-    // Note: CCTPv2 contracts must be deployed on both source and destination chains
-
-    // Arbitrum ↔ other EVM chains
-    {
-      src: '@Arbitrum',
-      dest: '@Ethereum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Ethereum',
-      dest: '@Arbitrum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Arbitrum',
-      dest: '@Base',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Base',
-      dest: '@Arbitrum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Arbitrum',
-      dest: '@Avalanche',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 30, // Avalanche has instant finality
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Avalanche',
-      dest: '@Arbitrum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Arbitrum',
-      dest: '@Optimism',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Optimism',
-      dest: '@Arbitrum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-
-    // Base ↔ other EVM chains (excluding Arbitrum, already defined)
-    {
-      src: '@Base',
-      dest: '@Ethereum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Ethereum',
-      dest: '@Base',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Base',
-      dest: '@Avalanche',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 30, // Avalanche has instant finality
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Avalanche',
-      dest: '@Base',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Base',
-      dest: '@Optimism',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Optimism',
-      dest: '@Base',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-
-    // Ethereum ↔ other EVM chains (excluding Arbitrum, Base already defined)
-    {
-      src: '@Ethereum',
-      dest: '@Avalanche',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 30, // Avalanche has instant finality
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Avalanche',
-      dest: '@Ethereum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Ethereum',
-      dest: '@Optimism',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Optimism',
-      dest: '@Ethereum',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-
-    // Avalanche ↔ Optimism (remaining pair)
-    {
-      src: '@Avalanche',
-      dest: '@Optimism',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 60,
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-    {
-      src: '@Optimism',
-      dest: '@Avalanche',
-      transfer: 'cctpV2',
-      variableFeeBps: 0,
-      timeSec: 30, // Avalanche has instant finality
-      min: 100_000n,
-      feeMode: 'evmToEvm',
-    },
-  ],
-});
-
+export { PROD_NETWORK };
 export default PROD_NETWORK;
