+++ packages/portfolio-contract/tools/target-balances.ts
@@ -1,0 +1,6 @@
+export {
+  computeTargetBalances,
+  DEFAULT_DELTA_SOFT_MIN,
+  TargetBalanceError,
+} from '@agoric/portfolio-api/src/target-balances.js';
+export type { ComputeTargetBalancesOptions } from '@agoric/portfolio-api/src/target-balances.js';
