+++ services/ymax-planner/src/main.ts
@@ -9,7 +9,7 @@ import * as ws from 'ws';
 
 import { Fail, q } from '@endo/errors';
 
-import { PROD_NETWORK } from '@aglocal/portfolio-contract/tools/network/prod-network.js';
+import { PROD_NETWORK } from '@agoric/portfolio-api/src/network/prod-network.js';
 import {
   fetchEnvNetworkConfig,
   getInvocationUpdate,
